requirements.txt
# Discord Bot
discord.py
watchdog
asyncpg

# Web Dashboard
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
